Add generic usePlayer for UI - video not working yet
