+++ src/Releases/Release0002_YearUnknown/Canvas.js
@@ -3,12 +3,7 @@ import useAudioPlayer from '../../UI/Player/hooks/useAudioPlayer';
 import '../Release.css';
 import Scene from './Scene';
 
-
 export default function Canvas({ }) {
   const { isPlaying } = useAudioPlayer();
   return <Scene paused={!isPlaying} />
 }
-
-
-
-
