+++ src/Releases/Release0008_GreemJellyfish/Canvas.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import useVideoPlayer from '../../UI/Player/hooks/useVideoPlayer';
+import '../Release.css';
+import Scene from './Scene';
+
+export default function Canvas({ }) {
+  const { videoPlayer } = useVideoPlayer();
+  return <Scene videoPlayer={videoPlayer} />
+}
