+++ src/UI/Player/hooks/usePlayer.js
@@ -1,0 +1,9 @@
+import useVideoPlayer from "./useVideoPlayer"
+import useAudioPlayer from "./useAudioPlayer";
+
+const usePlayer = (mediaType) => {
+    if (mediaType == "video") return useVideoPlayer();
+    else return useAudioPlayer();
+}
+
+export default usePlayer;
