+++ src/UI/UI.js
@@ -5,8 +5,6 @@ import Navigation from './Navigation';
 import Overlay from './Overlay/Overlay';
 import Player from './Player/Player';
 import './UI.css';
-import useAudioPlayer from './Player/hooks/useAudioPlayer';
-
 
 export default function UI({
     content,
